Solve requested changes

+++ src/main/java/com/autopartner/controller/rest/CompanyController.java
@@ -1,47 +1,50 @@
 package com.autopartner.controller.rest;
 
-import com.autopartner.DTO.CompanyRegistrationRequest;
+import com.autopartner.controller.DTO.CompanyRegistrationRequest;
 import com.autopartner.domain.Company;
-import com.autopartner.domain.User;
-import com.autopartner.service.impl.CompanyServiceImpl;
-import com.autopartner.service.impl.UserServiceImpl;
+import com.autopartner.service.CompanyService;
+import com.autopartner.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
-import org.springframework.stereotype.Controller;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 
+import java.util.NoSuchElementException;
+
 import static lombok.AccessLevel.PRIVATE;
 
-@Controller
+@RestController
+@RequestMapping("/api/company")
 @RequiredArgsConstructor
 @FieldDefaults(level = PRIVATE, makeFinal = true)
 public class CompanyController {
 
-    CompanyServiceImpl companyService;
-    UserServiceImpl userService;
+  CompanyService companyService;
+  UserService userService;
 
-    @Secured({"ROLE_ADMIN", "ROLE_ROOT"})
-    @GetMapping(value = "/api/company")
-    public ResponseEntity<?> getAll() {
-        return ResponseEntity.ok(companyService.listAllCompanies());
+  @Secured({"ROLE_ADMIN", "ROLE_ROOT"})
+  @GetMapping()
+  public Iterable<Company> getAll() {
+    return companyService.listAllCompanies();
+  }
+
+  @PostMapping(value = "/registration")
+  public CompanyRegistrationRequest newCompany(@Valid @RequestBody CompanyRegistrationRequest companyRegistrationRequest, BindingResult result) {
+    companyService.createCompany(companyRegistrationRequest);
+      return companyRegistrationRequest;
+  }
+
+  @Secured("ROLE_ADMIN")
+  @DeleteMapping(value = "/{id}")
+  public void deleteCompany(@PathVariable Long id) {
+    if (companyService.getCompanyById(id) != null) {
+      companyService.deleteCompany(id);
+    } else {
+      throw new NoSuchElementException("Company does not exist");
     }
-
-    @PostMapping(value = "/api/company/registration")
-    public ResponseEntity<?> newCompany(@Valid @RequestBody CompanyRegistrationRequest companyRegistrationRequest) {
-        companyService.saveCompany(Company.createCompany(companyRegistrationRequest));
-        userService.saveUser(User.createUser(companyRegistrationRequest));
-        return ResponseEntity.ok(companyRegistrationRequest);
-    }
-
-    @Secured("ROLE_ADMIN")
-    @DeleteMapping(value = "/api/company/{id}")
-    public void deleteCompany(@PathVariable Long id) {
-        companyService.deleteCompany(id);
-    }
-
+  }
 
 }
